requirements.txt
absl-py==0.13.0
aiohttp==3.7.4.post0
albumentations==1.0.3
antlr4-python3-runtime==4.8
appdirs==1.4.4
appnope==0.1.2
argon2-cffi==20.1.0
astunparse==1.6.3
async-generator==1.10
async-timeout==3.0.1
attrs==21.2.0
backcall==0.2.0
black==21.7b0
bleach==4.0.0
cachetools==4.2.2
certifi==2021.5.30
cffi==1.14.6
chardet==4.0.0
charset-normalizer==2.0.4
clang==5.0
click==8.0.1
configparser==5.0.2
cycler==0.10.0
debugpy==1.4.1
decorator==5.0.9
defusedxml==0.7.1
docker-pycreds==0.4.0
efficientnet-pytorch==0.7.1
entrypoints==0.3
flake8==3.9.2
flatbuffers==1.12
fsspec==2021.7.0
future==0.18.2
gast==0.4.0
gitdb==4.0.7
GitPython==3.1.18
google-auth==1.34.0
google-auth-oauthlib==0.4.5
google-pasta==0.2.0
graphviz==0.17
grpcio==1.39.0
h5py==3.1.0
hydra-core==1.1.0
idna==3.2
imageio==2.9.0
importlib-resources==5.2.2
ipykernel==6.0.3
ipython==7.26.0
ipython-genutils==0.2.0
ipywidgets==7.6.3
isort==5.9.3
jedi==0.18.0
Jinja2==3.0.1
joblib==1.0.1
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.12
jupyter-console==6.4.0
jupyter-contrib-core==0.3.3
jupyter-contrib-nbextensions==0.5.1
jupyter-core==4.7.1
jupyter-highlight-selected-word==0.2.0
jupyter-latex-envs==1.4.6
jupyter-nbextensions-configurator==0.4.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
jupyterthemes==0.20.0
keras==2.6.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
lesscpy==0.15.0
lightgbm==3.3.0
lxml==4.6.3
Markdown==3.3.4
MarkupSafe==2.0.1
matplotlib==3.4.2
matplotlib-inline==0.1.2
mccabe==0.6.1
mistune==0.8.4
multidict==5.1.0
mypy==0.910
mypy-extensions==0.4.3
nbclient==0.5.3
nbconvert==6.1.0
nbformat==5.1.3
nest-asyncio==1.5.1
networkx==2.6.2
notebook==6.4.2
numpy==1.19.5
oauthlib==3.1.1
omegaconf==2.1.0
opencv-contrib-python==4.5.3.56
opencv-python==4.5.3.56
opencv-python-headless==4.5.3.56
opt-einsum==3.3.0
packaging==21.0
pandas==1.3.1
pandocfilters==1.4.3
parso==0.8.2
pathspec==0.9.0
pathtools==0.1.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.3.1
ply==3.11
prometheus-client==0.11.0
promise==2.3
prompt-toolkit==3.0.19
protobuf==3.17.3
psutil==5.8.0
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.7.0
pycparser==2.20
pyDeprecate==0.3.1
pydotplus==2.0.2
pyflakes==2.3.1
Pygments==2.9.0
pyparsing==2.4.7
PyQt5==5.15.4
PyQt5-Qt5==5.15.2
PyQt5-sip==12.9.0
pyrsistent==0.18.0
python-dateutil==2.8.2
python-dotenv==0.19.0
pytorch-lightning==1.4.1
pytz==2021.1
PyWavelets==1.1.1
PyYAML==5.4.1
pyzmq==22.2.1
qtconsole==5.1.1
QtPy==1.9.0
regex==2021.8.3
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
scikit-image==0.18.2
scikit-learn==0.24.2
scipy==1.7.1
seaborn==0.11.1
Send2Trash==1.8.0
sentry-sdk==1.3.1
shortuuid==1.0.1
six==1.15.0
smmap==4.0.0
subprocess32==3.5.4
tensorboard==2.6.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow-addons==0.14.0
tensorflow-estimator==2.6.0
tensorflow-gpu==2.6.0
termcolor==1.1.0
terminado==0.10.1
testpath==0.5.0
threadpoolctl==2.2.0
tifffile==2021.8.8
timm==0.4.12
toml==0.10.2
tomli==1.2.1
torch==1.9.0+cu111
torchaudio==0.9.0
torchinfo==1.5.3
torchmetrics==0.5.0
torchvision==0.10.0+cu111
tornado==6.1
tqdm==4.62.0
traitlets==5.0.5
typeguard==2.13.0
typing-extensions==3.7.4.3
urllib3==1.26.6
wandb==0.12.0
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.0.1
widgetsnbextension==3.5.1
wrapt==1.12.1
yarl==1.6.3
zipp==3.5.0
